+++ apps/openmw-mp/Inventory.hpp
@@ -25,7 +25,7 @@ public:
 
     //inventory
     int getChangesSize() const;
-    void addItem(const std::string& refId, unsigned int count, int charge, double enchantmentCharge);
+    void addItem(const std::string& refId, unsigned int count, int charge, float enchantmentCharge);
     void removeItem(const std::string& refId, unsigned short count);
 
     /**
@@ -37,7 +37,7 @@ public:
 
 
     // equipment
-    void equipItem(unsigned short slot, const std::string& refId, unsigned int count, int charge, double enchantmentCharge);
+    void equipItem(unsigned short slot, const std::string& refId, unsigned int count, int charge, float enchantmentCharge);
     void unequipItem(unsigned short slot);
 
     bool hasItemEquipped(const std::string& refId) const;
